separate out logger and run_control into sep files
