+++ pyam_analysis/logger.py
@@ -1,0 +1,14 @@
+import logging
+
+# globally accessible logger
+_LOGGER = None
+
+
+def logger():
+    """Access global logger"""
+    global _LOGGER
+    if _LOGGER is None:
+        logging.basicConfig()
+        _LOGGER = logging.getLogger()
+        _LOGGER.setLevel('INFO')
+    return _LOGGER
